Move the communication to the SOAP client
